third_party/blink/web_tests/platform/generic/fast/js/pic/get-empty-string-expected.txt

This page tests the empty string used as a property name. If the test passes, you'll see a PASS message below.

PASS: o[""] should be 1 and is.
